# api-docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "NFC Loyalty & Customer Growth Platform API",
    "description": "REST API untuk platform loyalty berbasis NFC/QR untuk UMKM Indonesia. Menyediakan fitur registrasi member, pemesanan, validasi pembayaran manual, poin loyalitas, voucher, dan dashboard analytics.",
    "version": "1.0.0",
    "contact": { "name": "NFC Loyalty Platform" }
  },
  "servers": [
    { "url": "http://localhost:3000", "description": "Development server" }
  ],
...
  "paths": {
...
    "/api/auth/register": {
      "post": {
        "tags": ["Auth"],
        "summary": "Register new member",
        "description": "Register a new member. If WhatsApp already exists, performs auto-login. Issues welcome voucher if configured.",
        "operationId": "registerMember",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RegisterRequest" },
              "example": {
                "name": "John Doe",
                "whatsapp": "081234567890",
                "tenantId": "550e8400-e29b-41d4-a716-446655440000",
                "accessMethod": "nfc"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Registration successful",
            "content": {
              "application/json": {
                "example": {
                  "success": true,
                  "data": {
                    "memberId": "a1b2c3d4-e5f6-7890-abcd-ef1234567890",
                    "accessToken": "eyJhbGciOiJIUzI1NiJ9.eyJtZW1iZXJJZCI6ImExYjJjM2Q0In0.abc123",
                    "refreshToken": "eyJhbGciOiJIUzI1NiJ9.eyJtZW1iZXJJZCI6ImExYjJjM2Q0In0.refresh456",
                    "welcomeVoucher": {
                      "code": "WELCOME-XYZ789",
                      "discountType": "percentage",
                      "discountValue": 10,
                      "expiryDate": "2025-08-15T00:00:00.000Z"
                    }
                  }
                }
              }
            }
          },
          "200": {
            "description": "Member already exists - auto login",
            "content": {
              "application/json": {
                "example": {
                  "success": true,
                  "data": {
                    "memberId": "a1b2c3d4-e5f6-7890-abcd-ef1234567890",
                    "accessToken": "eyJhbGciOiJIUzI1NiJ9.access_token_here",
                    "refreshToken": "eyJhbGciOiJIUzI1NiJ9.refresh_token_here"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "example": {
                  "success": false,
                  "error": {
                    "message": "Validation failed",
                    "details": [
                      { "field": "name", "message": "Nama minimal 2 karakter" },
                      { "field": "whatsapp", "message": "Format nomor WhatsApp tidak valid (08xx atau +628xx, 10-13 digit)" }
                    ]
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "tags": ["Auth"],
        "summary": "Member login by WhatsApp",
        "operationId": "loginMember",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/LoginRequest" },
              "example": {
                "whatsapp": "081234567890",
                "tenantId": "550e8400-e29b-41d4-a716-446655440000"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful",
            "content": {
              "application/json": {
                "example": {
                  "success": true,
                  "data": {
                    "memberId": "a1b2c3d4-e5f6-7890-abcd-ef1234567890",
                    "accessToken": "eyJhbGciOiJIUzI1NiJ9.member_access",
                    "refreshToken": "eyJhbGciOiJIUzI1NiJ9.member_refresh"
                  }
                }
              }
            }
          },
          "404": {
            "description": "WhatsApp number not registered",
            "content": {
              "application/json": {
                "example": {
                  "success": false,
                  "error": { "message": "Nomor WhatsApp belum terdaftar", "code": "MEMBER_NOT_FOUND" }
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/refresh": {
      "post": {
        "tags": ["Auth"],
        "summary": "Refresh access token",
        "operationId": "refreshToken",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["refreshToken"],
                "properties": {
                  "refreshToken": { "type": "string" }
                }
              },
              "example": {
                "refreshToken": "eyJhbGciOiJIUzI1NiJ9.your_refresh_token_here"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Token refreshed",
            "content": {
              "application/json": {
                "example": {
                  "success": true,
                  "data": {
                    "accessToken": "eyJhbGciOiJIUzI1NiJ9.new_access_token",
                    "refreshToken": "eyJhbGciOiJIUzI1NiJ9.new_refresh_token"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Invalid or expired refresh token",
            "content": {
              "application/json": {
                "example": {
                  "success": false,
                  "error": { "message": "Invalid or expired refresh token" }
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/admin/login": {
      "post": {
        "tags": ["Auth"],
        "summary": "Admin login",
        "operationId": "adminLogin",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/AdminLoginRequest" },
              "example": {
                "email": "admin@tokoku.com",
                "password": "password123"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Admin login successful",
            "content": {
              "application/json": {
                "example": {
                  "success": true,
                  "data": {
                    "adminId": "b2c3d4e5-f6a7-8901-bcde-f12345678901",
                    "accessToken": "eyJhbGciOiJIUzI1NiJ9.admin_access_token",
                    "refreshToken": "eyJhbGciOiJIUzI1NiJ9.admin_refresh_token"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials",
            "content": {
              "application/json": {
                "example": {
                  "success": false,
                  "error": { "message": "Email atau password salah" }
                }
              }
            }
          }
        }
      }
    },
    "/api/members/me": {
      "get": {
        "tags": ["Members"],
        "summary": "Get current member profile",
        "operationId": "getMemberProfile",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Member profile",
            "content": {
              "application/json": {
                "example": {
                  "success": true,
                  "data": {
                    "id": "a1b2c3d4-e5f6-7890-abcd-ef1234567890",
                    "tenantId": "550e8400-e29b-41d4-a716-446655440000",
                    "name": "John Doe",
                    "whatsapp": "081234567890",
                    "pointBalance": 150,
                    "totalVisits": 12,
                    "registeredAt": "2025-01-15T08:30:00.000Z",
                    "lastVisitAt": "2025-07-14T14:20:00.000Z"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "example": { "success": false, "error": { "message": "Member access required" } }
              }
            }
          }
        }
      },
      "patch": {
        "tags": ["Members"],
        "summary": "Update member name",
        "operationId": "updateMemberProfile",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name"],
                "properties": {
                  "name": { "type": "string", "minLength": 2, "maxLength": 50 }
                }
              },
              "example": { "name": "Budi Santoso" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Profile updated",
            "content": {
              "application/json": {
                "example": {
                  "success": true,
                  "data": {
                    "id": "a1b2c3d4-e5f6-7890-abcd-ef1234567890",
                    "tenantId": "550e8400-e29b-41d4-a716-446655440000",
                    "name": "Budi Santoso",
                    "whatsapp": "081234567890",
                    "pointBalance": 150,
                    "totalVisits": 12,
                    "registeredAt": "2025-01-15T08:30:00.000Z",
                    "lastVisitAt": "2025-07-14T14:20:00.000Z"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "example": { "success": false, "error": { "message": "Nama minimal 2 karakter" } }
              }
            }
          }
        }
      }
    },
    "/api/admin/members": {
      "get": {
        "tags": ["Members"],
        "summary": "List all members (admin)",
        "operationId": "listMembers",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 1 } },
          { "name": "pageSize", "in": "query", "schema": { "type": "integer", "default": 20, "maximum": 100 } }
        ],
        "responses": {
          "200": {
            "description": "Paginated member list",
            "content": {
              "application/json": {
                "example": {
                  "success": true,
                  "data": {
                    "members": [
                      {
                        "id": "a1b2c3d4-e5f6-7890-abcd-ef1234567890",
                        "name": "John Doe",
                        "whatsapp": "081234567890",
                        "pointBalance": 150,
                        "totalVisits": 12,
                        "registeredAt": "2025-01-15T08:30:00.000Z",
                        "lastVisitAt": "2025-07-14T14:20:00.000Z"
                      }
                    ],
                    "pagination": { "page": 1, "pageSize": 20, "totalItems": 45, "totalPages": 3 }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT access token (15min). Get from /api/auth/register, /api/auth/login, or /api/auth/admin/login"
      }
    },
    "schemas": {
      "RegisterRequest": {
        "type": "object",
        "required": ["name", "whatsapp", "tenantId", "accessMethod"],
        "properties": {
          "name": { "type": "string", "minLength": 2, "maxLength": 100, "description": "Letters, spaces, periods, apostrophes only" },
          "whatsapp": { "type": "string", "description": "Indonesian format: 08xx, +628xx, or 628xx (10-13 digits)" },
          "tenantId": { "type": "string", "format": "uuid" },
          "accessMethod": { "type": "string", "enum": ["nfc", "qr", "direct"] }
        }
      },
      "LoginRequest": {
        "type": "object",
        "required": ["whatsapp", "tenantId"],
        "properties": {
          "whatsapp": { "type": "string" },
          "tenantId": { "type": "string", "format": "uuid" }
        }
      },
      "AdminLoginRequest": {
        "type": "object",
        "required": ["email", "password"],
        "properties": {
          "email": { "type": "string", "format": "email" },
          "password": { "type": "string" }
        }
      },
...
    }
  }
}